Remove KNN9120Codec, update tests
